[Store] Extract MetadataBackend interface for pluggable metadata backends
Refactor MasterService to delegate all metadata operations to a backend
abstraction layer, enabling future support for alternative metadata stores
(e.g., Redis) without modifying the RPC interface.

Changes:
- Add MetadataBackend interface with 67 pure virtual methods mirroring
  MasterService RPC surface (ExistKey, GetReplicaList, Put, MountSegment, etc.)
- Create NativeMetadataBackend class that contains the original in-memory
  metadata logic extracted from MasterService
- Refactor MasterService to delegate all operations to backend_ member
- Extract SegmentDetailInfo to segment.h and add UUIDHash to types.h for
  cleaner interface boundaries

This is Phase 1 of 4 in the Redis backend implementation plan. Pure refactor
with zero behavior change - MasterService still provides the same RPC interface,
just delegates to NativeMetadataBackend internally.

Next phases will add RedisMetadataBackend as an alternative implementation.
