+++ mooncake-store/src/CMakeLists.txt
@@ -4,6 +4,7 @@ add_subdirectory(cachelib_memory_allocator)
 set(MOONCAKE_STORE_SOURCES
     allocator.cpp
     master_service.cpp
+    native_metadata_backend.cpp
     client_service.cpp
     client_metric.cpp
     types.cpp
